moved namespaces filter to Resources Section
